Hi 서비스 분리 리팩터링

+++ src/main/java/msg/team1/Hi/domain/notice/presentation/dto/response/GetIdNoticeResponse.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor(access = AccessLevel.PROTECTED) @AllArgsConstructor
 public class GetIdNoticeResponse {
 
-    private Integer noticeId;
+    private Long noticeId;
     private String title;
     private String content;
     private String name;
@@ -20,5 +20,7 @@ public class GetIdNoticeResponse {
 
     @LastModifiedDate
     private LocalDateTime modifiedDate;
+
+
 
 }
